Move timer definitions to submodule
